Demo 07 - Spy and mock
Here we can now isolate a lot more dependencies, and we can control our tests better.

One of the biggest improvements here is that we no longer need to post the inventory to be able to test how our component works with different inventory levels: we can mock the inventory levels.